HST/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
appdirs==1.4.4
    # via fs
astropy==5.3.4
    # via
    #   astroquery
    #   pyvo
astroquery==0.4.6
    # via -r requirements.in
beautifulsoup4==4.12.2
    # via astroquery
build==1.0.3
    # via pip-tools
certifi==2023.7.22
    # via requests
cffi==1.16.0
    # via xattr
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
coverage==7.3.2
    # via -r requirements.in
exceptiongroup==1.1.3
    # via pytest
fs==2.4.16
    # via -r requirements.in
html5lib==1.1
    # via astroquery
idna==3.4
    # via requests
importlib-metadata==6.8.0
    # via
    #   build
    #   keyring
iniconfig==2.0.0
    # via pytest
jaraco-classes==3.3.0
    # via keyring
keyring==24.2.0
    # via astroquery
more-itertools==10.1.0
    # via jaraco-classes
numpy==1.26.1
    # via
    #   astropy
    #   astroquery
    #   pyerfa
    #   rms-julian
    #   rms-pdslogger
    #   rms-textkernel
packaging==23.2
    # via
    #   astropy
    #   build
    #   pytest
pip-tools==7.3.0
    # via -r requirements.in
pluggy==1.3.0
    # via pytest
pycparser==2.21
    # via cffi
pyerfa==2.0.1.1
    # via astropy
pyparsing==3.1.1
    # via
    #   -r requirements.in
    #   rms-julian
    #   rms-pdslogger
    #   rms-textkernel
pyproject-hooks==1.0.0
    # via build
pytest==7.4.3
    # via -r requirements.in
pyvo==1.4.2
    # via astroquery
pyyaml==6.0.1
    # via astropy
requests==2.31.0
    # via
    #   astroquery
    #   pyvo
rms-julian==2.0.0
    # via
    #   -r requirements.in
    #   rms-pdstemplate
    #   rms-textkernel
rms-pdslogger==0.0.5
    # via
    #   -r requirements.in
    #   rms-pdstemplate
rms-pdstemplate==0.0.3
    # via -r requirements.in
rms-textkernel==1.0.0
    # via
    #   rms-julian
    #   rms-pdslogger
six==1.16.0
    # via
    #   fs
    #   html5lib
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy==2.0.23
    # via -r requirements.in
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
    #   pytest
typing-extensions==4.8.0
    # via sqlalchemy
unidecode==1.3.7
    # via -r requirements.in
urllib3==2.0.7
    # via
    #   -r requirements.in
    #   requests
webencodings==0.5.1
    # via html5lib
wheel==0.41.3
    # via pip-tools
xattr==0.10.1
    # via rms-pdslogger
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
